Buggy Chess Rules and File Structure
This commit includes many more files related to creating the model subsystem of the game and linking it all. It's still highly buggy as it crashes easily and the chess pieces end up in the wrong spot. The filesystem makes more easier compililation and sorts cpp files and hpp files.